Add a bit of a better tool for handling services deps
Right now it's a function.  Eventually, it'll be an app
from out of this app.  It's clumsy in a way to just
use, "sv check" with a timeout to do this.  You need
better and something that shoves out good feedback.

So I added a useful scripting function, if even it
is a hack, to handle this functionality for all users.