Backing the console gettys support out for now.
Code passed testing but had a major regression not caught
out except in one of the other community member's target
builds.  Since I'm on a time critical path for them and
I needed that for the larger picture for them and my own
inscrutable projects, removing it to revisit with better
later.