Add concept of service groups proper..
- Clean up the options processing so it's consistent in all cases.
- Add the services group bbclass which adds a notion of services
  groupings for recipes where the runit runners are grouped in
  the recipe instead of when they're bbappended.

Ultimately, we want the generator to touch the "down" database
entries in the specified recipies regardless of whether the
runit database exists in the specified recipe.  We want to
extend the concept of services groups to a bigger picture.
For now, though, this will do and provide a good core for
people to use to make much more sophisticated init/supervision
configs.
  to re-work it so it can do down files into packaging for any
  named