Launch tuning.
There was a bit of start up consistency problems and overall
speed concerns that this addresses.  We wait 2 seconds for
the check to come back for the prior group to report start
or not.  We wait the default 7 for a launch (shouldn't
take that long, but...) and we used pause stupidly
in the finish call, which hangs some start sequences
making it not boot consistently.  (Sometimes things
would timeout on the check or fail- we want it to
go through the current sequencing.  Ideally, we
want the sleep to go away and have services behave
themselves better, but not right now...)