Fix logging config support.
Followed Gerrit's documentation rather than auditing the code
and verifying what it actually does.  In general, this is a sound
practice, but in this case, logging config is read out of the logdir
which while accurate on the smarden.org website and the internal
manpages, is very, very misleading.   Fixed wiki.  Will fix
internals here in a moment on runit.  Need to fix it here as well.
These changes reflect this functionality along with the fix for
a very clear incorrect sense of timestamping setup.