Additional clean-up moves for external
Our core is based off of meta, not meta-openembedded, meta-efl, etc.
so we are going to move them to an external layer corresponding to
the layer.conf naming of the layers in question so that when
you don't have a given meta-<foo> layer in your BBLAYERS it doesn't
get confused and complain about not having <bar>.bb for the
corresponding .bbappend that enables runit support for that daemon.