Silly me.  No LICENSE file.  Well then, let's FIX that, shall we?
For the record: Retroactively to it's beginnings when it landed
on GitHub, going forward, this has always been under MIT/X11.

In the large, metadata is pure functional and who "wrote" what
and the rights therein belongs to all parties involved, including
any corporations.  Hence MIT/X11 for this.  It just doesn't need
to be plagerized by anyone and maintained by everyone on behalf
of the community as a whole.